refactor k8s client
