+++ tests/integrationTests/yamlFiles/deployAppTest.yaml
@@ -89,8 +89,11 @@ metadata:
   name: token-review-cluster-role
 rules:
 - apiGroups: ["authentication.k8s.io"]
-  resources: ["tokenreviews","serviceaccounts"]
+  resources: ["tokenreviews"]
   verbs: ["create","watch","get","list"]
+- apiGroups: [""]
+  resources: ["serviceaccounts"]
+  verbs: ["watch","get","list"]
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
